Fixed seg fault caused by sytax error during polyInfo allocation

+++ C_wrappers.py
@@ -31,7 +31,7 @@ loadPolygons = c_funcs.loadPolygons
 loadPolygons.argtypes = [c_int,
                          c_int,
                          ndpointer(dtype=c_double, ndim=3, flags='C_CONTIGUOUS'),
-                         ndpointer(dtype=c_double, ndim=1, flags='C_CONTIGUOUS')]
+                         ndpointer(dtype=c_int, ndim=1, flags='C_CONTIGUOUS')]
 
 ## SETUP I2C ##
 setupI2C = c_funcs.setupI2C
